fix service of order and model [error -> int - long]

+++ src/main/java/blc/idat/apiblc/service/PedidoService.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface PedidoService {
 
-    List<Pedido> findbyCliente(int id);
+    List<Pedido> findbyCliente(Long id);
 }
